V0.2
Change GUNINATOR and weapon behaviour
Added assets and prefab
Fix AI issue
Created display scene

+++ Assets/Assets/Scripts/Weapon_1.0/Bullet/BulletStats_ForEnemy.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class BulletStats_ForEnemy : MonoBehaviour
 {
     public float bullet_Speed;
-    public int bullet_Damage;
-    public int round_Type;
+    public float bullet_Damage;
+    public float round_Type;
     public float debuff_Effect_Time;
 
     Rigidbody the_RB;
